refactor: add minor changes
