+++ src/main/java/ru/practicum/shareit/booking/model/State.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.booking;
+package ru.practicum.shareit.booking.model;
 
 public enum State {
     ALL,
