+++ src/main/java/ru/practicum/shareit/request/model/ItemRequest.java
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 import ru.practicum.shareit.user.model.User;
 
 import javax.persistence.*;
-import java.time.LocalDate;
 
 @Data
 @Builder
